Units: added "fixed" methods
New wrappers which work with "fixed" instead of "double".  When the
transition to "fixed" is completed, we will remove the "double"
methods.

+++ src/Dialogs/dlgBasicSettings.cpp
@@ -117,8 +117,8 @@ OnAltitudeData(DataField *Sender, DataField::DataAccessKind_t Mode)
 {
   switch (Mode) {
   case DataField::daGet:
-    Sender->Set(Units::ToUserUnit(XCSoarInterface::Basic().BaroAltitude,
-                                  Units::AltitudeUnit));
+    Sender->Set((double)Units::ToUserUnit(XCSoarInterface::Basic().BaroAltitude,
+                                          Units::AltitudeUnit));
     break;
   case DataField::daPut:
     break;
